Move SciSpacyAnnotator to its own module for siloing purposes
